Create Button_Function.pde
When using this function, type "button(a,b,c,d,e,f)", where a= the base colour of the button: typed with #, followed by 6 capital letters (or just use the colour selector in tools), b= the x value of the top left corner of the box, c= the y value of the top left corner of the box, d= the x value of the bottom right corner of the box, e= the y value of the bottom right corner of the box, and f= the colour the button changes to when its hovered over: typed with #, followed by 6 capital letters (or just use the colour selector in tools). Ask me if you're too dumb to understand that.